requirements.txt
pandas==2.2.2
Pillow==10.3.0
streamlit==1.32.0
streamlit_pandas_profiling==0.1.3
ydata_profiling==4.8.3
